# docs/swagger.yaml
basePath: /
definitions:
  access.UpdateExpiredDateRequest:
    properties:
      expired_date:
        type: string
    type: object
  access.UpdateRateLimitRequest:
    properties:
      rate_limit:
        minimum: 1
        type: integer
    required:
    - rate_limit
    type: object
  access.User:
    properties:
      email:
        type: string
      expired_date:
        type: string
      group:
        $ref: '#/definitions/group.Group'
      group_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      rate_limit:
        description: Requests per minute
        type: integer
      status_id:
        type: integer
    type: object
  audit.AuditLog:
    properties:
      api_key:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      method:
        type: string
      path:
        type: string
      request_body:
        type: string
      request_headers:
        type: string
      response_body:
        type: string
      response_time:
        description: in milliseconds
        type: integer
      status_code:
        type: integer
      status_id:
        type: integer
      user_agent:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
    type: object
...
  group.Group:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/permission.Permission'
        type: array
      status_id:
        type: integer
    type: object
...
  permission.Permission:
    properties:
      action:
        description: e.g., "create", "read", "update", "delete"
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        description: e.g., "examples", "users", "reports"
        type: string
      status_id:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    name: API Support
  description: This is a modular REST API built with Go Fiber
  title: My API
  version: "1.0"
paths:
  /v1/access/{id}/expired-date:
    delete:
      consumes:
      - application/json
      description: Remove the expiration date for a user's API key (never expires)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove API key expiration date
      tags:
      - Access
    put:
      consumes:
      - application/json
      description: Update the expiration date for a user's API key
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Expiration date data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/access.UpdateExpiredDateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update API key expiration date
      tags:
      - Access
  /v1/access/{id}/rate-limit:
    put:
      consumes:
      - application/json
      description: Update the rate limit for a user's API key (requests per minute)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rate limit data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/access.UpdateRateLimitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update API key rate limit
      tags:
      - Access
  /v1/audit-logs:
    get:
      consumes:
      - application/json
      description: Get audit logs with filtering and pagination
      parameters:
      - description: Filter by user email
        in: query
        name: user_email
        type: string
      - description: Filter by HTTP method
        in: query
        name: method
        type: string
      - description: Filter by API path
        in: query
        name: path
        type: string
      - description: Filter by status code
        in: query
        name: status_code
        type: integer
      - description: Filter from date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter to date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: 'Limit results (default: 50, max: 1000)'
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get audit logs
      tags:
      - Audit
  /v1/audit-logs/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed audit log by ID including request/response body
      parameters:
      - description: Audit Log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audit.AuditLog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log by ID
      tags:
      - Audit
  /v1/audit-logs/cleanup:
    delete:
      consumes:
      - application/json
      description: Delete audit logs older than specified days
      parameters:
      - description: Delete logs older than this many days
        in: query
        name: days
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete old audit logs
      tags:
      - Audit
...
  /v1/profile:
    get:
      consumes:
      - application/json
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/access.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Access
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your API token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"